Method to change cleaning mode remotely
Code has been implemented to change the cleaning operation on the fly. To do this, you need to be in remote mode, then hold down all of the buttons for a little over 5 seconds. Then for the next 5 seconds (a little over), choose a particular combination of inputs to choose a particular cleaning operation. Once that time has elapsed, you can now try out your new cleaning operation,