Remote single button
Now there is a choice on how a cleaning mode can be chosen in remote mode. Instead of using the directional inputs, a single button can be wired to REMOTE PRESET port on the board. Now the user has full control on how they wish to change cleaning modes in remote  mode.